Moved crosshair narrators to the API
