+++ common/src/main/java/org/mcaccess/minecraftaccess/addon/CoreAddon.java
@@ -1,5 +1,6 @@
 package org.mcaccess.minecraftaccess.addon;
 
+import dev.architectury.platform.Platform;
 import net.minecraft.client.Minecraft;
 import net.minecraft.tags.FluidTags;
 import org.jetbrains.annotations.NotNull;
@@ -14,6 +15,8 @@ import org.mcaccess.minecraftaccess.addon.accessmenu.NarrateTarget;
 import org.mcaccess.minecraftaccess.addon.accessmenu.OpenConfig;
 import org.mcaccess.minecraftaccess.addon.accessmenu.RefreshScreenReader;
 import org.mcaccess.minecraftaccess.addon.accessmenu.TargetPosition;
+import org.mcaccess.minecraftaccess.addon.crosshairnarrators.Jade;
+import org.mcaccess.minecraftaccess.addon.crosshairnarrators.MCAccess;
 import org.mcaccess.minecraftaccess.addon.statuses.Air;
 import org.mcaccess.minecraftaccess.addon.statuses.Armour;
 import org.mcaccess.minecraftaccess.addon.statuses.Durability;
@@ -90,5 +93,10 @@ public class CoreAddon implements MinecraftAccessAddon {
         registry.register("xp", new GetXP());
         registry.register("refresh_screen_reader", new RefreshScreenReader());
         registry.register("config", new OpenConfig());
+
+        registry.register("minecraft_access", new MCAccess());
+        if (Platform.isModLoaded("jade")) {
+            registry.register("jade", new Jade());
+        }
     }
 }
