+++ common/src/main/resources/assets/minecraft_access/lang/en_us.json
@@ -276,6 +276,8 @@
     "minecraft_access.wolf_variant.woods": "Woods",
     "minecraft_access.xp_indicator.decreased": "Player XP level decreased to %s",
     "minecraft_access.xp_indicator.increased": "Player XP level increased to %s",
+    "narrator.minecraft_access.jade": "Jade",
+    "narrator.minecraft_access.minecraft_access": "Minecraft Access",
     "text.autoconfig.minecraft-access.category.accessMenu": "Access Menu",
     "text.autoconfig.minecraft-access.category.cameraControls": "Camera",
     "text.autoconfig.minecraft-access.category.default": "General",
@@ -339,12 +341,12 @@
     "text.autoconfig.minecraft-access.option.narrateCrosshair.filter.whitelist": "Whitelist",
     "text.autoconfig.minecraft-access.option.narrateCrosshair.narrateAdditionalEntityPoses": "Narrate additional entity poses",
     "text.autoconfig.minecraft-access.option.narrateCrosshair.narrateBlockFace": "Narrate block face",
+    "text.autoconfig.minecraft-access.option.narrateCrosshair.narrator": "Narrator",
     "text.autoconfig.minecraft-access.option.narrateCrosshair.relativePositionSoundCue": "Relative position sound cue",
     "text.autoconfig.minecraft-access.option.narrateCrosshair.relativePositionSoundCue.enabled": "Enabled",
     "text.autoconfig.minecraft-access.option.narrateCrosshair.relativePositionSoundCue.maxSoundVolume": "Maximum volume",
     "text.autoconfig.minecraft-access.option.narrateCrosshair.relativePositionSoundCue.minSoundVolume": "Minimum volume",
     "text.autoconfig.minecraft-access.option.narrateCrosshair.repetitionInterval": "Repetition interval",
-    "text.autoconfig.minecraft-access.option.narrateCrosshair.useJade": "Use Jade",
     "text.autoconfig.minecraft-access.option.playerWarnings.airThreshold": "Air threshold",
     "text.autoconfig.minecraft-access.option.playerWarnings.durabilityWarnings": "Durability Warnings",
     "text.autoconfig.minecraft-access.option.playerWarnings.durabilityWarnings.enableHeldItems": "Enable held item warnings",
